+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## TBD
+  - `Default` implementation
+
 ## v0.1 (02-06-2022)
   - documentation, readme, CI
   - tests with Loom and Miri